Report parsing errors with Esprima
Esprima gives us much better parsing errors for certain syntax problems,
e.g.:

```javascript
var "foo";
"foo" = "bar";
```
