+++ locales/en.json
@@ -78,6 +78,12 @@
       "invalid-tag-name": "<__tag__> isn’t a valid HTML tag. If you want to create a custom tag, make sure the name has a - in it, like <my-tag>"
     },
     "javascript": {
+      "unexpected-string": "JavaScript doesn’t expect a string where you wrote __value__. Remember that variable and function names aren’t surrounded in quotes.",
+      "unexpected-number": "JavaScript doesn’t expect a number where you wrote __value__. Remember that variable and function names have to start with a letter.",
+      "invalid-left-hand-string": "You wrote __value__ to the left of the = sign here, but that’s not a valid variable name. Remember that variable names aren’t surrounded by quotes.",
+      "invalid-left-hand-number": "You wrote __value__ to the left of the = sign here, but that’s not a valid variable name. Remember that variable names have to start with a letter.",
+      "invalid-left-hand-boolean": "You wrote __value__ to the left of the = sign here, but that’s not a valid variable name. Remember that true and false can’t be used as variable names.",
+      "invalid-left-hand-value": "You wrote __value__ to the left of the = sign here, but that’s not a valid variable name.",
       "unmatched": "There is a __openingSymbol__ on this line that needs a closing __closingSymbol__",
       "closing-match": "There is a __openingSymbol__ on this line that needs a closing __closingSymbol__",
       "expected-identifier": "There's something missing from this line. Did you forget something?",
